requirements.txt
numpy>=1.26,<2.1
pandas>=2.2,<2.3
scipy>=1.11,<1.13
statsmodels>=0.14,<0.15
lifelines>=0.27,<0.28
matplotlib>=3.7,<3.9
seaborn>=0.13,<0.14
PyYAML>=6.0,<7
patsy>=0.5,<0.6
typer[all]>=0.12,<0.13
pandera>=0.17,<0.18
scikit-learn>=1.3,<1.5
tqdm>=4.66,<5
plotly>=5.22,<5.23
# Optional: BigQuery access (not used in demo run)
google-cloud-bigquery>=3.14,<4
pandas-gbq>=0.21,<0.22
pyarrow>=17,<18
